+++ cmd/sample-api/main.go
@@ -5,17 +5,17 @@ import (
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
-	"github.com/podhmo-sandbox/sample-api/controller"
 	"github.com/podhmo-sandbox/sample-api/model/repository"
+	"github.com/podhmo-sandbox/sample-api/webapi/todo"
 )
 
 func mount(r chi.Router) {
 	r.Route("/todos", func(r chi.Router) {
 		repo := repository.NewTodoRepository()
-		r.MethodFunc("GET", "", controller.GetTodos(repo))
-		r.MethodFunc("POST", "", controller.PostTodo(repo))
-		r.MethodFunc("PUT", "", controller.PutTodo(repo))
-		r.MethodFunc("DELETE", "", controller.DeleteTodo(repo))
+		r.MethodFunc("GET", "", todo.GetTodos(repo))
+		r.MethodFunc("POST", "", todo.PostTodo(repo))
+		r.MethodFunc("PUT", "", todo.PutTodo(repo))
+		r.MethodFunc("DELETE", "", todo.DeleteTodo(repo))
 	})
 }
